refactor: provide current conversation in context provider
