+++ web/src/components/SideMenu/index.jsx
@@ -16,7 +16,7 @@ import ChatTab from "components/SideMenu/SideMenuButton";
  */
 const SideMenu = () => {
   const [username,] = useContext(UserContext);
-  const [conversations, dispatch] = useContext(ConversationContext);
+  const [conversations, , dispatch] = useContext(ConversationContext);
 
   const createChat = () => {
     createConversation()
